deps/v8/test/inspector/debugger/wasm-regress-10957-expected.txt

Regress 10957

Running test: testRegress10957
Instantiate
7
